Major reorg of Impala SQL translation toolchain and e2e tests
