+++ setup.py
@@ -115,10 +115,18 @@ setup(
     packages=['ibis',
               'ibis.expr',
               'ibis.expr.tests',
+              'ibis.hive',
+              'ibis.hive.tests',
               'ibis.impala',
               'ibis.impala.tests',
               'ibis.sql',
               'ibis.sql.tests',
+              'ibis.sql.presto',
+              'ibis.sql.presto.tests',
+              'ibis.sql.redshift',
+              'ibis.sql.redshift.tests',
+              'ibis.sql.vertica',
+              'ibis.sql.vertica.tests',
               'ibis.tests'],
     version=VERSION,
     package_data={'ibis': ['*.pxd', '*.pyx']},
